chore: rename to remove-empty-fields for consistency
