+++ src/utils/remove-empty-fields.js
